requirements.txt
Django==1.9.7
djangorestframework
markdown
django-filter
MySQL-Python
jsonpickle
django-cors-headers
djangorestframework-jsonp
